Add unit tests for setupFlags in separate file
Moved the tests for the setupFlags function from config/config_test.go to a new file, config/flags_test.go. This helps to separate concerns and keeps the config tests file more focused.
